feat:  - add internal transfer operation  - add get wallet balance  - add get wallet movements

+++ src/main/kotlin/com/iobuilders/bank/poc/domain/Money.kt
@@ -1,3 +1,5 @@
 package com.iobuilders.bank.poc.domain
 
-data class Money(var amount: Float, val currency: MoneyCurrency)
+import java.math.BigDecimal
+
+data class Money(var amount: BigDecimal, val currency: MoneyCurrency)
